+++ addOns/requester/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Add a context menu to open the Manual Request Editor, on ZAP versions newer than 2.11.
+
 ### Changed
 - Maintenance changes.